+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [0.17.0] - 2024-12-17
 
 - Fix sort option label being empty
 - Add a summary of the search query above the results (optional, disabled by
@@ -220,6 +220,7 @@ The minimum Omeka S version required is 3.0.0
 
 First release
 
+[0.17.0]: https://github.com/biblibre/omeka-s-module-Search/releases/tag/v0.17.0
 [0.16.0]: https://github.com/biblibre/omeka-s-module-Search/releases/tag/v0.16.0
 [0.15.5]: https://github.com/biblibre/omeka-s-module-Search/releases/tag/v0.15.5
 [0.15.4]: https://github.com/biblibre/omeka-s-module-Search/releases/tag/v0.15.4